requirements.txt
onnxruntime>=1.17.0
omegaconf>=2.3.0
wandb>=0.16.3
torchmetrics>=1.3.0
onnx>=1.15.0
tqdm>=4.66.1